requirements.txt
numpy<2
Pillow
torch
torchvision
flask

